app/pyproject.toml
[tool.poetry]
name = "mot-eggs-app"
version = "v0.0.1"
description = "Multiple object tracking of easter eggs in an app"
authors = ["Martin Højland Hansen (hojland93@gmail.com)"]

[tool.poetry.dependencies]
python = "3.9.0"
pandas = "^1.1.4"
pydantic = "^1.7.2"
python-dotenv = "^0.15.0"
tqdm = "^4.60.0"
fastapi = "^0.75.0"
jmespath = "^0.10.0"
loguru = "^0.5.3"
python-multipart = "^0.0.5"
google-cloud-logging = "^2.6.0"
httpx = {extras = ["http2"], version = "^0.23.0"}
sentry-sdk = "^1.5.12"
PyJWT = "^2.3.0"
cryptography = "^36.0.2"
toml = "^0.10.2"
gradio = "^3.19.1"
torch = "^1.13.1"
GitPython = "^3.1.31"
Pillow = "^9.0.0"
filterpy = "^1.4.5"
scikit-image = "^0.19.2"
matplotlib = "^3.2.2"
requests = "^2.23.0"
scipy = "^1.4.1"
Cython = "^0.29.28"
numpy = ">=1.0.0,<1.20.0"
torchvision = "^0.14.1"
thop = "^0.1.1.post2209072238"
pycocotools = "^2.0.6"
seaborn = "^0.12.2"
opencv-python = "4.5.2.54"
videoio = "^0.2.8"
av = "^10.0.0"
ffmpeg-python = "^0.2.0"
poethepoet = "^0.18.1"

[tool.poetry.dependencies.uvicorn]
version = ">=0.18.3"
extras = [ "standard" ]

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"
ipykernel = "^5.3.4"
ipywidgets = "^7.6.3"
jupyterlab = "^2.2.9"
mypy = "^0.770"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest-black = "^0.3.10"
pytest-clarity = "^0.3.0-alpha.0"
pytest-mock = "^3.3.1"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.11.1"
Faker = "^4.1.1"
google-cloud-storage = "^1.41.1"

[tool.poetry.dev-dependencies.black]
version = "^22.1"
allow-prereleases = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--black"
testpaths = "tests"

[tool.poe.tasks]
install_lap = "pip3 install lap cython-bbox"

[tool.black]
line-length = 135
target-version = [ "py38" ]
include = "\\.pyi?$"
exclude = """
/(
    \\.eggs         # exclude a few common directories in the
  | \\.git          # root of the project
  | \\.mypy_cache
  | \\.venv
  | \\.vscode
  | _build
  | buck-out
  | build
  | dist
  | tmp
)/
"""

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[build-system]
requires = [ "poetry>=1.1.8" ]
build-backend = "poetry.masonry.api"
